Separate playbook for generating certificates.
When running on the VM’s a prerequisite will be to generate the
various certificates (for SP’s, TLS, etc) in a separate step.
This reduces the number of environment specific roles.

The playbook openconext-generate-certs-vm.yml will
place all generated certificates and keys in the files/{java|php}-vm
folders. The various roles for the “normal” playbooks assume these
files to be present.