Fix tomcat setup.
The name of the xml file in the deploy directory (e.g. ROOT.xml) matters
since that should be the same name as the context path. So is the context
path is / than the xml file should be called ROOT.xml.
If the context path is /teams than the name of the xml must be teams.xml.