Dev-up setup, leveraging bash scripting, ansible tags and vagrant env variables
Added skip_develop tag to Ansible that can be used to skip certain steps in development, leveraging the --skip-tags flag when running ./ansible-env.
Added an ENV variable that can be used by Vagrant to allow mounting of directories when ENV=develop.
Added a provision-develop bash file that starts Vagrant with ENV=develop and Ansible provision-vm.yml with develop as environment.
Added a --reload flag to provision-develop so Vagrant can be reloaded with the development environment without reprovisioning (./provision-develop --reload)