Disable Haproxy health checks to speed up development
The haproxy health checks meant that a backend server
was considered down when a development error led to a
400 or 500 range status code. This can be rather cumbersome
since the healthchecks are configured to take some time
inter 20000 fall 2 rise 2 -> each 40 seconds which means
that if a server is down due to development it can take
up to 40 seconds to get back to life, extending the
feedback cycle. By making the stats socket in development
admin level, we're allowed to send the 'disable health'
command for backends. By sending these commands for all
backends, the health status check is ignored, allowing
easier development.