Add engineblock base_domain to configuration for Symfony
The Symfony routing component routes based on host name
(amongst other params), this is dependent on the base
domain the services run on. In order to be able to
redirect to routes, we need to have the domain so URLs
can be generated. This can be done by using a default
based on a parameter. This parameter is set in the
kernel and loaded from the EngineBlock configuration.