Make default NameIDFormat transient for new entities.
When we add a new entity we want to default to releasing as little information
as possible. Also, when we release persistent nameid we don't know whether the
SP will use it. If we release a transient one, we can be sure they don't use
it to uniquely identify users.