Moving the restart to the springbooapp tasks
If it was the first deployment of pdp, the service state=started would
wait until the service was started and the next thing was that the
'restart pdp' handler was fired. This is prevented by only using the
handler. Creating a new symlink (in the springbootapp tasks) will also
fire the restart handler