Add SIRTFI Security contact boolean isSirtfiSecurityContact.
SIRTFI provides for a way to put incident response contact info into
metadata. The contactType used is "other" which we already support.
However, this contact is then enriched with other attributes that
indicates it's SIRTFI. So we need a boolean to specify this. The type
should be set to "other" when the boolean is set.

Also allow one more contact to be specified per entity as the 3
contacts are normally already in use.

See also: https://refeds.org/sirtfi