Give vagrant user all rights to engineblock files
This is required to be able to run composer in the development
environment without having to resort to using sudo. This is
necessary due to the files and directories being owned by the
engineblock fpm user. This means a composer process invoked by
the vagrant user may not alter the files/dirs. Previously this was
resolved by using sudo, however composer blocked being run with
root privileges as a security measure. By giving full rights,
using sudo is no longer necessary.