Remove Janus references from eb.ini file
This is part of the cleanup that was performed in
OpenConext-engineblock. When building a new EngineBlock development env,
EB will fail on invalid configuration that is provided in the two ini
files. The application.ini file (override) was updated in the EB repo.
The one written to /etc/openconext was yet to be updated.

And that is taken care of in this PR.

See https://github.com/OpenConext/OpenConext-engineblock/pull/581