Use the PHP 7.2 xdebug extension
Currently we use the PHP 5.6 xdebug extension which does not work for
PHP 7.2. By creating an additional xdebug ini file for the PHP 7.2 pool,
with a link to the correct extension, the debug feature should now work
as intended on both PHP 5.6 and 7.2 applications.

See: https://github.com/OpenConext/OpenConext-deploy/pull/233#issuecomment-532594993