Unconditionally ensure GUI link is correct
Changing the GUI symlink was dependent on something getting unarchived.
When the deploy failed on a step after unrarchive, and you re-run it when
fixed, the GUI symlink is not changed because unarchive was already done.
Remove the condition, so always is ensured that the symlink points
to the correct place.