Provision manage entity names also in nl and pt variants (using en string)
This provides a better default initial experience if your browser is
set to any of the other supported languages than en. Otherwise EB
will show the entityID of the IdP in the WAYF, for example, by
lack of a name:nl. Just fill it with name_en for now, not a lot of
need to provide actual translations of these names for an initial
filling.