Fix: push_mw_* tags not working
The push_mw_* tags used to push new configuration to the middleware stopped working.
The problem is that "import_tasks" is not run when one the push_mw_* tags is set, so the docker/vm tasks are not included.
Fixed by explicitly adding the tags to the import_tasks condition when so the tasks are imported.